requirements.txt
aiosmtplib==2.0.1
alembic==1.10.4
anyio==3.6.2
async-generator==1.10
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.95.1
fastapi-dependency==0.1.0
fastapi-mail==1.2.8
fastapi-pagination==0.12.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
outcome==1.2.0
packaging==23.1
passlib==1.7.4
patch==1.16
pluggy==1.2.0
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PySocks==1.7.1
pytest==7.3.2
pytest-asyncio==0.21.0
pytest-mock==3.11.1
pytest-tornasync==0.6.0.post2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.13
sqlalchemy-pagination==0.0.2
starlette==0.26.1
tomli==2.0.1
tornado==6.3.2
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.5.0
urllib3==2.0.3
uvicorn==0.22.0
wsproto==1.2.0